dont use cargo workspaces
apparently there is a bug/misfeature which prevents us from setting
the panic strategy for a specific package in the workspace, which
is broken because then running `cargo check` from the workspace root
fails, because it doesn't take into account the panic strategy of the
inner packages.

instead just use seperate crates which are in a shared folder.

+++ kernel/loader/src/main.rs
@@ -111,7 +111,7 @@ fn read_cp0_config() -> u32 {
 }
 
 fn initialize_cache() {
-    let mmu_type: usize = read_cp0_config();
+    let mmu_type: u32 = read_cp0_config();
 }
 
 global_asm!(
